+++ Sources/MapboxMaps/Gestures/GestureOptions.swift
@@ -24,9 +24,9 @@ public struct GestureOptions: Equatable {
     /// Defaults to `true`.
     public var pinchEnabled: Bool = true
     
-    /// Whether the pinch gesture rotation is enabled.
+    /// Whether rotation is enabled for the pinch gesture.
     /// Defaults to `true`.
-    public var pinchRotationEnabled: Bool = true
+    public var pinchRotateEnabled: Bool = true
 
     /// Whether the pitch gesture is enabled. Defaults to `true`.
     public var pitchEnabled: Bool = true