+++ Tests/MapboxMapsTests/Gestures/GestureHandlers/Mocks/MockPinchGestureHandler.swift
@@ -1,5 +1,5 @@
 @testable import MapboxMaps
 
 final class MockPinchGestureHandler: GestureHandler, PinchGestureHandlerProtocol {
-    var rotationEnabled: Bool = true
+    var rotateEnabled: Bool = true
 }